pyproject.toml
[project]
name = "officeboxpy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
dependencies = [
    "flet-core>=0.24.1",
    "flet-pages>=0.0.2",
    "flet-runtime>=0.24.1",
    "pyinstaller>=6.16.0",
]

[tool.flet.macos]
entitlement."com.apple.security.personal-information.photos-library" = true
